create the queryKeys file
In this change, I created a queryKeys.ts file to export the querykeys as objects and also deleted some unnecssary packages
