+++ pages/queryKeys.ts
@@ -1,0 +1,5 @@
+const queryKeys = {
+  COMMENT: "comment",
+};
+
+export { queryKeys };
